+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#1877](https://github.com/clj-kondo/clj-kondo/issues/1876): suppress `new` analysis from `(String. x)` expansion
 - [#1888](https://github.com/clj-kondo/clj-kondo/issues/1876): use `namespace-munge` for resolving hook files rather than `munge`
 - [#1896](https://github.com/clj-kondo/clj-kondo/issues/1896): don't consider `**`, `***` etc. to be a dynamic vars
+- [#1899](https://github.com/clj-kondo/clj-kondo/issues/1899): treat var or local reference as unused value when not in tail position
 
 ## 2022.11.02